requirements.txt
numpy>=1.19.5
opencv-python>=4.5.4.60
ultralytics>=8.0.0
Pillow>=9.0.0
cryptography>=36.0.0
